# docs/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp AI Agent Platform API",
    "description": "Production-grade SaaS WhatsApp AI Agent Platform with Multi-Agent & Graph RAG architecture",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@whatsappagent.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/reminders": {
      "get": {
        "tags": ["Reminders"],
        "summary": "List reminders",
        "description": "Get all reminders for the authenticated user",
        "operationId": "listReminders",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["pending", "completed", "cancelled"]
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start date filter",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date filter",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminders retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminders": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Reminder" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reminders"],
        "summary": "Create reminder",
        "description": "Create a new reminder",
        "operationId": "createReminder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reminder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reminder": { "$ref": "#/components/schemas/Reminder" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/search": {
      "get": {
        "tags": ["Documents"],
        "summary": "Search documents",
        "description": "Search documents by query",
        "operationId": "searchDocuments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Search query"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": { "type": "array", "items": { "type": "object" } },
                        "query": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload document",
        "description": "Upload a new document for processing",
        "operationId": "uploadDocument",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "document": { "$ref": "#/components/schemas/Document" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Get document",
        "description": "Get document details",
        "operationId": "getDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Document retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "document": { "$ref": "#/components/schemas/Document" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "summary": "Delete document",
        "description": "Delete a document",
        "operationId": "deleteDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{id}/process": {
      "post": {
        "tags": ["Documents"],
        "summary": "Process document",
        "description": "Start document processing",
        "operationId": "processDocument",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Processing started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string" }
            }
          }
        }
      },
...
      "Reminder": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "scheduledAt": { "type": "string", "format": "date-time" },
          "timezone": { "type": "string" },
          "repeatType": { "type": "string", "enum": ["NONE", "DAILY", "WEEKLY", "MONTHLY", "YEARLY"] },
          "repeatInterval": { "type": "integer" },
          "repeatEndDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["PENDING", "COMPLETED", "CANCELLED"] },
          "metadata": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateReminderRequest": {
        "type": "object",
        "required": ["title", "scheduledAt"],
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "scheduledAt": { "type": "string", "format": "date-time" },
          "timezone": { "type": "string", "default": "UTC" },
          "repeatType": { "type": "string", "enum": ["none", "daily", "weekly", "monthly", "yearly"], "default": "none" },
          "repeatInterval": { "type": "integer", "minimum": 1, "default": 1 },
          "repeatEndDate": { "type": "string", "format": "date-time" },
          "metadata": { "type": "object" }
        }
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "fileType": { "type": "string" },
          "fileSize": { "type": "integer" },
          "status": { "type": "string", "enum": ["uploading", "processing", "indexed", "failed"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UploadDocumentRequest": {
        "type": "object",
        "required": ["title", "fileType", "fileSize"],
        "properties": {
          "title": { "type": "string" },
          "fileType": { "type": "string" },
          "fileSize": { "type": "integer" }
        }
      },
...
    }
  }
}