# services/api-gateway/src/routes/document.ts
import { Router } from 'express';
import { z } from 'zod';
import { authenticate, AuthenticatedRequest } from '../middleware/auth';

const router = Router();

router.get('/search', authenticate, async (req: AuthenticatedRequest, res, next) => {
  try {
    const { q, limit } = req.query;
    
    res.json({
      success: true,
      data: {
        results: [],
        query: q,
      },
    });
  } catch (error) {
    next(error);
  }
});

router.post('/upload', authenticate, async (req: AuthenticatedRequest, res, next) => {
  try {
    const { title, fileType, fileSize } = req.body;
    
    res.json({
      success: true,
      data: {
        document: {
          id: 'doc_xxx',
          title,
          fileType,
          fileSize,
          status: 'processing',
        },
      },
    });
  } catch (error) {
    next(error);
  }
});

router.get('/:id', authenticate, async (req: AuthenticatedRequest, res, next) => {
  try {
    res.json({
      success: true,
      data: {
        document: {
          id: req.params.id,
          title: 'Document',
          status: 'indexed',
        },
      },
    });
  } catch (error) {
    next(error);
  }
});

router.delete('/:id', authenticate, async (req: AuthenticatedRequest, res, next) => {
  try {
    res.json({
      success: true,
      data: { message: 'Document deleted' },
    });
  } catch (error) {
    next(error);
  }
});

router.post('/:id/process', authenticate, async (req: AuthenticatedRequest, res, next) => {
  try {
    res.json({
      success: true,
      data: { message: 'Document processing started' },
    });
  } catch (error) {
    next(error);
  }
});

export { router };